Convert Clarg to .Net Core tooling. Modify project to target .Net 4.5.1 and .Net Core Platform 1.2. Clean up some whitespace. Fix a reflection call that doesn't work in .Net Core. Bump version to 2.0.1.
