+++ Clarg.sln
@@ -1,0 +1,33 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 14
+VisualStudioVersion = 14.0.25123.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Clarg", "src\Clarg\Clarg.xproj", "{6AEAA003-8F25-481A-B9E2-46D40EB623A7}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{E1355C49-9CF9-4A2B-AAB8-B353D1F6C75C}"
+	ProjectSection(SolutionItems) = preProject
+		global.json = global.json
+	EndProjectSection
+EndProject
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Clarg.Test", "test\Clarg.Test\Clarg.Test.xproj", "{914AD591-4011-406C-AAF9-8514712BBBC9}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{6AEAA003-8F25-481A-B9E2-46D40EB623A7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6AEAA003-8F25-481A-B9E2-46D40EB623A7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6AEAA003-8F25-481A-B9E2-46D40EB623A7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6AEAA003-8F25-481A-B9E2-46D40EB623A7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{914AD591-4011-406C-AAF9-8514712BBBC9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{914AD591-4011-406C-AAF9-8514712BBBC9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{914AD591-4011-406C-AAF9-8514712BBBC9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{914AD591-4011-406C-AAF9-8514712BBBC9}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
