+++ global.json
@@ -1,0 +1,8 @@
+{
+	"projects": [ "src", "test" ],
+	"sdk": {
+		"version": "1.0.0-rc1-update2",
+		"runtime": "coreclr",
+		"architecture": "x64"
+	}
+}
