--- lib/.placeholder
@@ -1,1 +1,0 @@
-This is just a placeholder so Git will create the folder.
