--- src/Clarg.sln
@@ -1,28 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2013
-VisualStudioVersion = 12.0.30501.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Clarg", "Clarg\Clarg.csproj", "{D326332C-4D6A-4378-90C9-2DD15937919F}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tests", "Tests\Tests.csproj", "{05966093-97BB-4A66-AA51-57D30BBA1051}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{D326332C-4D6A-4378-90C9-2DD15937919F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D326332C-4D6A-4378-90C9-2DD15937919F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D326332C-4D6A-4378-90C9-2DD15937919F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D326332C-4D6A-4378-90C9-2DD15937919F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{05966093-97BB-4A66-AA51-57D30BBA1051}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{05966093-97BB-4A66-AA51-57D30BBA1051}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{05966093-97BB-4A66-AA51-57D30BBA1051}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{05966093-97BB-4A66-AA51-57D30BBA1051}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
