+++ src/Clarg/ArgumentDescriptor.cs
@@ -17,10 +17,10 @@ namespace Clarg
 
 		public override bool Equals(object obj)
 		{
-			if(Object.ReferenceEquals(obj, null))
+			if(ReferenceEquals(obj, null))
 				return false;
 
-			if(Object.ReferenceEquals(obj, this))
+			if(ReferenceEquals(obj, this))
 				return true;
 
 			if(obj is ArgumentDescriptor)
@@ -31,10 +31,10 @@ namespace Clarg
 
 		public bool Equals(ArgumentDescriptor obj)
 		{
-			if(Object.ReferenceEquals(obj, null))
+			if(ReferenceEquals(obj, null))
 				return false;
 
-			if(Object.ReferenceEquals(obj, this))
+			if(ReferenceEquals(obj, this))
 				return true;
 
 			return StringComparer.OrdinalIgnoreCase.Equals(obj.Name, Name) && StringComparer.OrdinalIgnoreCase.Equals(obj.Value, Value);
@@ -47,10 +47,10 @@ namespace Clarg
 
 		public static bool operator ==(ArgumentDescriptor x, ArgumentDescriptor y)
 		{
-			if(Object.ReferenceEquals(x, null) || Object.ReferenceEquals(y, null))
-				return Object.ReferenceEquals(x, null) && Object.ReferenceEquals(y, null);
+			if(ReferenceEquals(x, null) || ReferenceEquals(y, null))
+				return ReferenceEquals(x, null) && ReferenceEquals(y, null);
 
-			if(Object.ReferenceEquals(x, y))
+			if(ReferenceEquals(x, y))
 				return true;
 
 			return x.Equals(y);
