+++ src/Clarg/ArgumentParameterMapping.cs
@@ -17,10 +17,10 @@ namespace Clarg
 
 		public override bool Equals(object obj)
 		{
-			if(Object.ReferenceEquals(obj, null))
+			if(ReferenceEquals(obj, null))
 				return false;
 
-			if(Object.ReferenceEquals(obj, this))
+			if(ReferenceEquals(obj, this))
 				return true;
 
 			if(obj is ArgumentParameterMapping)
@@ -31,10 +31,10 @@ namespace Clarg
 
 		public bool Equals(ArgumentParameterMapping obj)
 		{
-			if(Object.ReferenceEquals(obj, null))
+			if(ReferenceEquals(obj, null))
 				return false;
 
-			if(Object.ReferenceEquals(obj, this))
+			if(ReferenceEquals(obj, this))
 				return true;
 
 			return Argument == obj.Argument && Parameter == obj.Parameter;
@@ -49,10 +49,10 @@ namespace Clarg
 
 		public static bool operator ==(ArgumentParameterMapping x, ArgumentParameterMapping y)
 		{
-			if(Object.ReferenceEquals(x, null) || Object.ReferenceEquals(y, null))
-				return Object.ReferenceEquals(x, null) && Object.ReferenceEquals(y, null);
+			if(ReferenceEquals(x, null) || ReferenceEquals(y, null))
+				return ReferenceEquals(x, null) && ReferenceEquals(y, null);
 
-			if(Object.ReferenceEquals(x, y))
+			if(ReferenceEquals(x, y))
 				return true;
 
 			return x.Equals(y);
