+++ src/Clarg/project.json
@@ -1,0 +1,25 @@
+{
+	"version": "2.0.1-*",
+	"description": "Command Line Arguments Parser",
+	"authors": [ "jason@vortx.com", "josh@vortx.com" ],
+	"projectUrl": "https://github.com/Vortx-Inc/Clarg",
+	"licenseUrl": "",
+
+	"frameworks": {
+		"net451": {
+			"dependencies": {
+				"System.Diagnostics.Debug": "4.0.11-beta-23516"
+			}
+		},
+		"dotnet5.2": {
+			"dependencies": {
+				"Microsoft.CSharp": "4.0.0",
+				"System.ComponentModel.TypeConverter": "4.0.0",
+				"System.Linq": "4.0.0",
+				"System.Reflection": "4.0.0",
+				"System.Runtime": "4.0.10",
+				"System.Runtime.Extensions": "4.0.0"
+			}
+		}
+	}
+}
