+++ test/Clarg.Test/Properties/AssemblyInfo.cs
@@ -1,0 +1,9 @@
+﻿using System.Reflection;
+
+[assembly: AssemblyTitle("Clarg.Tests")]
+[assembly: AssemblyDescription("Clarg test suite")]
+[assembly: AssemblyCompany("Vortx, Inc.")]
+[assembly: AssemblyProduct("Clarg")]
+[assembly: AssemblyCopyright("Copyright © Vortx, Inc. 2016")]
+
+[assembly: AssemblyVersion("2.0.1.*")]
