+++ test/Clarg.Test/project.json
@@ -1,0 +1,29 @@
+{
+	"version": "2.0.1-*",
+	"description": "Clarg test suite",
+	"authors": [ "jason@vortx.com", "josh@vortx.com" ],
+	"projectUrl": "https://github.com/Vortx-Inc/Clarg",
+	"licenseUrl": "",
+	"frameworks": {
+		"dnx451": {
+			"frameworkAssemblies": {
+				"System.Runtime": "4.0.0.0"
+			}
+		},
+		"dnxcore50": {
+			"dependencies": {
+				"Microsoft.CSharp": "4.0.0",
+				"System.Linq": "4.0.0",
+				"System.Runtime": "4.0.10"
+			}
+		}
+	},
+	"dependencies": {
+		"xunit": "2.1.0",
+		"xunit.runner.dnx": "2.1.0-rc1-build204",
+		"Clarg": "2.0.1-*"
+	},
+	"commands": {
+		"test": "xunit.runner.dnx"
+	}
+}
