--- test/TestRunner.msbuild
@@ -1,19 +1,0 @@
-<Project
-    DefaultTargets="Test"
-    xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-
-	<UsingTask
-		AssemblyFile="..\lib\packages\xunit.runners.1.9.2\tools\xunit.runner.msbuild.dll"
-		TaskName="Xunit.Runner.MSBuild.xunit" />
-
-	<Target Name="Test">
-		<ItemGroup>
-			<TestAssemblies Condition="Exists('bin\Test.dll')" Include="bin\Test.dll" />
-			<TestAssemblies Condition="Exists('bin\Tests.dll')" Include="bin\Tests.dll" />
-			<TestAssemblies Include="bin\*.Test.dll" />
-		</ItemGroup>
-
-		<xunit Condition="Exists('@(TestAssemblies)')" Assemblies="@(TestAssemblies)" />
-	</Target>
-
-</Project>
